Refactor PDF generation
